+++ frontend_tests/node_tests/dispatch_subs.js
@@ -187,7 +187,6 @@ test("stream create", (override) => {
     const stub = make_stub();
     override(stream_data, "create_streams", stub.f);
     override(stream_data, "get_sub_by_id", noop);
-    override(stream_data, "update_calculated_fields", noop);
     override(subs, "add_sub_to_table", noop);
     override(overlays, "streams_open", () => true);
     dispatch(event);
